@lucianaTSoares [Task: Validação checklist]

+++ src/main/java/org/generation/blogpessoal/controller/PostController.java
@@ -1,6 +1,9 @@
 package org.generation.blogpessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.generation.blogpessoal.model.Post;
 import org.generation.blogpessoal.repository.PostRepository;
@@ -16,6 +19,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
 @RequestMapping("/postagens")
@@ -31,27 +35,34 @@ public class PostController {
     }
 
     @GetMapping("/{id_post}")
-    public ResponseEntity<Post> getById(@PathVariable Integer id_post) {
-        return repository.findById(id_post).map(resp -> ResponseEntity.ok(resp)).orElse(ResponseEntity.notFound().build());
+    public ResponseEntity<Post> getById(@PathVariable Long id_post) {
+        return repository.findById(id_post).map(resp -> ResponseEntity.ok(resp))
+        .orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
     }
 
     @GetMapping("/titulo/{titulo}")
-    public ResponseEntity<List<Post>> getByTitle(@PathVariable String titulo) {
+    public ResponseEntity<List<Post>> getByTitulo(@PathVariable String titulo) {
         return ResponseEntity.ok(repository.findAllByTituloContainingIgnoreCase(titulo));
     }
 
     @PostMapping
-    public ResponseEntity<Post> post(@RequestBody Post postagem) {
+    public ResponseEntity<Post> post(@Valid @RequestBody Post postagem) {
         return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(postagem));
     }
 
     @PutMapping
-    public ResponseEntity<Post> put(@RequestBody Post postagem) {
-        return ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem));
+    public ResponseEntity<Post> put(@Valid @RequestBody Post postagem) {
+        return repository.findById(postagem.getId_post()).map(resp -> ResponseEntity.status(HttpStatus.OK)
+        .body(repository.save(postagem))).orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
     }
 
     @DeleteMapping("/{id_post}")
-    private void delete(@PathVariable Integer id_post) {
+    private void delete(@PathVariable Long id_post) {
+        Optional<Post> post = repository.findById(id_post);
+
+        if (post.isEmpty())
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Postagem não encontrada.", null);
+
         repository.deleteById(id_post);
     }
 }
